+++ src/functions/index.ts
@@ -11,8 +11,6 @@ import {
 import {
   EventTopicBackfillForAll,
   EventTopicCleanBackfillForAll,
-  EventTopicGenerateCodeForJsonSchema,
-  EventTopicGenerateCodeReferencedInProjectForAll,
   EventTopicListForAll,
   EventTopicListReferencedInProjectForServerlessFunctions,
   EventTopicListReferencedInProjectForServiceContainer,
@@ -46,8 +44,6 @@ export function registerFunctions(context: ModuleRegistrationContext) {
     EnvironmentPrepareForAll,
     EventTopicBackfillForAll,
     EventTopicCleanBackfillForAll,
-    EventTopicGenerateCodeForJsonSchema,
-    EventTopicGenerateCodeReferencedInProjectForAll,
     EventTopicListReferencedInProjectForServerlessFunctions,
     EventTopicListReferencedInProjectForServiceContainer,
     EventTopicListForAll,